+++ dockstore-integration-testing/src/test/java/io/dockstore/webservice/ServiceIT.java
@@ -1,55 +1,60 @@
 /*
- * Copyright 2023 OICR and UCSC
+ *    Copyright 2019 OICR
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
  *
- *           http://www.apache.org/licenses/LICENSE-2.0
+ *        http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
  */
-
 package io.dockstore.webservice;
 
-import static io.dockstore.webservice.Constants.LAMBDA_FAILURE;
-import static io.dockstore.webservice.helpers.GitHubAppHelper.DockstoreTestUser2Repos;
-import static io.dockstore.webservice.helpers.GitHubAppHelper.handleGitHubRelease;
+import static io.dockstore.webservice.Constants.DOCKSTORE_YML_PATH;
+import static io.dockstore.webservice.resources.ResourceConstants.PAGINATION_LIMIT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNotSame;
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.dockstore.client.cli.BaseIT;
+import io.dockstore.client.cli.BaseIT.TestStatus;
 import io.dockstore.common.CommonTestUtilities;
 import io.dockstore.common.ConfidentialTest;
+import io.dockstore.common.DescriptorLanguage;
 import io.dockstore.common.MuteForSuccessfulTests;
 import io.dockstore.common.SourceControl;
-import io.dockstore.openapi.client.ApiClient;
-import io.dockstore.openapi.client.ApiException;
-import io.dockstore.openapi.client.api.UsersApi;
-import io.dockstore.openapi.client.api.WorkflowsApi;
-import io.dockstore.openapi.client.model.Workflow;
-import io.dockstore.openapi.client.model.WorkflowSubClass;
-import io.dockstore.webservice.core.EntryTypeMetadata;
-import io.dockstore.webservice.core.SourceFile;
+import io.dockstore.webservice.core.BioWorkflow;
+import io.dockstore.webservice.core.Service;
+import io.dockstore.webservice.core.User;
+import io.dockstore.webservice.core.Workflow;
+import io.dockstore.webservice.core.WorkflowMode;
 import io.dockstore.webservice.jdbi.FileDAO;
 import io.dockstore.webservice.jdbi.ServiceDAO;
 import io.dockstore.webservice.jdbi.UserDAO;
 import io.dockstore.webservice.jdbi.WorkflowDAO;
+import io.dropwizard.client.JerseyClientBuilder;
+import io.swagger.client.ApiClient;
 import io.swagger.client.api.Ga4GhApi;
+import io.swagger.client.api.WorkflowsApi;
+import io.swagger.client.model.StarRequest;
 import io.swagger.client.model.Tool;
+import jakarta.ws.rs.client.Client;
+import java.util.HashSet;
 import java.util.List;
-import org.apache.http.HttpStatus;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 import org.hibernate.context.internal.ManagedSessionContext;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -58,11 +63,15 @@ import uk.org.webcompere.systemstubs.jupiter.SystemStubsExtension;
 import uk.org.webcompere.systemstubs.stream.SystemErr;
 import uk.org.webcompere.systemstubs.stream.SystemOut;
 
+/**
+ * @author dyuen
+ */
 @ExtendWith(SystemStubsExtension.class)
 @ExtendWith(MuteForSuccessfulTests.class)
-@ExtendWith(BaseIT.TestStatus.class)
+@ExtendWith(TestStatus.class)
 @Tag(ConfidentialTest.NAME)
-public class ServiceIT extends BaseIT {
+class SwaggerServiceIT extends BaseIT {
+
     @SystemStub
     public final SystemOut systemOut = new SystemOut();
     @SystemStub
@@ -94,165 +103,200 @@ public class ServiceIT extends BaseIT {
         ManagedSessionContext.bind(session);
     }
 
-    @BeforeEach
-    @Override
-    public void resetDBBetweenTests() {
-        CommonTestUtilities.cleanStatePrivate2(SUPPORT, false, testingPostgres);
+    @Test
+    void checkWorkflowAndServiceHierarchy() {
+        CreateContent createContent = new CreateContent().invoke(false);
+        long workflowID = createContent.getWorkflowID();
+        long serviceID = createContent.getServiceID();
+        long serviceID2 = createContent.getServiceID2();
+
+        // might not be right if our test database is larger than PAGINATION_LIMIT
+        final List<Workflow> allPublished = workflowDAO.findAllPublished(0, Integer.valueOf(PAGINATION_LIMIT), null, null, null);
+        assertTrue(allPublished.stream().anyMatch(workflow -> workflow.getId() == workflowID && workflow instanceof BioWorkflow));
+        assertTrue(allPublished.stream().anyMatch(workflow -> workflow.getId() == serviceID && workflow instanceof Service));
+        assertTrue(allPublished.stream().anyMatch(workflow -> workflow.getId() == serviceID2 && workflow instanceof Service));
+
+        final Service byId = serviceDAO.findById(serviceID);
+        final Service byId1 = serviceDAO.findById(workflowID);
+
+        assertTrue(byId != null && byId1 == null);
+        session.close();
     }
 
-    /**
-     * This tests endpoints that will be triggered by GitHub App webhooks.
-     * A service is created and a version is added for a release 1.0
-     */
     @Test
-    void testGitHubAppEndpoints() {
-        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi client = new WorkflowsApi(webClient);
+    void checkServiceInCollection() {
 
-        // Add version
-        handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/1.0", USER_2_USERNAME);
-        long workflowCount = testingPostgres.runSelectStatement("select count(*) from service", long.class);
-        assertEquals(1, workflowCount);
-        Workflow service = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_SERVICE, WorkflowSubClass.SERVICE, "versions");
+        final io.dockstore.openapi.client.ApiClient webClientAdminUser = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
+        final io.dockstore.openapi.client.api.EntriesApi entriesApi = new io.dockstore.openapi.client.api.EntriesApi(webClientAdminUser);
+        final io.dockstore.openapi.client.api.OrganizationsApi organizationsApi = new io.dockstore.openapi.client.api.OrganizationsApi(webClientAdminUser);
 
-        assertNotNull(service);
-        assertEquals(1, service.getWorkflowVersions().size(), "Should have a new version");
-        List<SourceFile> sourceFiles = fileDAO.findSourceFilesByVersion(service.getWorkflowVersions().get(0).getId());
-        assertEquals(3, sourceFiles.size(), "Should have 3 source files");
+        io.dockstore.openapi.client.model.Organization organization = new io.dockstore.openapi.client.model.Organization();
+        organization.setName("serviceOrg");
+        organization.setDisplayName("serviceOrg");
+        organization.setEmail("test@email.com");
+        organization.setDescription("service service service");
+        organization.setTopic("This is a short topic");
+        io.dockstore.openapi.client.model.Organization collectionOrg =  organizationsApi.createOrganization(organization);
 
-        long users = testingPostgres.runSelectStatement("select count(*) from user_entry where entryid = '" + service.getId() + "'", long.class);
-        assertEquals(1, users, "Should have 1 user");
+        //approve organizations
+        collectionOrg = organizationsApi.approveOrganization(collectionOrg.getId());
 
-        final long count = testingPostgres.runSelectStatement(
-            "select count(*) from service where sourcecontrol = 'github.com' and organization = 'DockstoreTestUser2' and repository = 'test-service'",
-            long.class);
-        assertEquals(1, count, "there should be one matching service");
+        //create collection
+        io.dockstore.openapi.client.model.Collection collection = new io.dockstore.openapi.client.model.Collection();
+        collection.setName("Collection");
+        collection.setDisplayName("Collection");
+        collection.setDescription("A collection of notebooks");
+        collection = organizationsApi.createCollection(collection, collectionOrg.getId());
 
-        // Test user endpoints
-        UsersApi usersApi = new UsersApi(webClient);
-        final long userId = testingPostgres.runSelectStatement("select userid from user_entry where entryid = '" + service.getId() + "'", long.class);
-        List<Workflow> services = usersApi.userServices(userId);
-        List<Workflow> workflows = usersApi.userWorkflows(userId);
-        assertEquals(1, services.size(), "There should be one service");
-        assertEquals(0, workflows.size(), "There should be no workflows");
+        CreateContent createContent = new CreateContent().invoke(false);
+        long serviceID = createContent.getServiceID();
 
-        // Should not be able to refresh service
-        ApiException ex = assertThrows(ApiException.class, () -> client.refresh1(services.get(0).getId(), false));
-        assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode(), "Should fail since you cannot refresh services.");
+        //add service to collection
+        Set<String> expectedCollectionNames = new HashSet<>();
+        expectedCollectionNames.add("Collection");
+        organizationsApi.addEntryToCollection(collectionOrg.getId(), collection.getId(), serviceID, null);
+        List<io.dockstore.openapi.client.model.CollectionOrganization> entryCollection = entriesApi.entryCollections(serviceID);
+        assertEquals(expectedCollectionNames,  entryCollection.stream().map(io.dockstore.openapi.client.model.CollectionOrganization::getCollectionName).collect(Collectors.toSet()));
+        assertEquals(1, entryCollection.stream().map(io.dockstore.openapi.client.model.CollectionOrganization::getCollectionName).collect(Collectors.toSet()).size());
+        assertEquals(0, organizationsApi.getCollectionByName(collectionOrg.getName(), collection.getName()).getWorkflowsLength());
+        assertEquals(1, organizationsApi.getCollectionByName(collectionOrg.getName(), collection.getName()).getServicesLength());
+
+        //remove service from collection
+        organizationsApi.deleteEntryFromCollection(collectionOrg.getId(), collection.getId(), serviceID, null);
+        expectedCollectionNames.remove("Collection");
+        entryCollection = entriesApi.entryCollections(serviceID);
+        assertEquals(expectedCollectionNames,  entryCollection.stream().map(io.dockstore.openapi.client.model.CollectionOrganization::getCollectionName).collect(Collectors.toSet()));
+        assertEquals(0, entryCollection.stream().map(io.dockstore.openapi.client.model.CollectionOrganization::getCollectionName).collect(Collectors.toSet()).size());
+        assertEquals(0, organizationsApi.getCollectionByName(collectionOrg.getName(), collection.getName()).getWorkflowsLength());
+        assertEquals(0, organizationsApi.getCollectionByName(collectionOrg.getName(), collection.getName()).getServicesLength());
     }
 
-    /**
-     * Ensures that you cannot create a service if the given user is not on Dockstore
-     */
     @Test
-    void createServiceNoUser() {
-        final ApiClient webClient = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
-        WorkflowsApi client = new WorkflowsApi(webClient);
-
-        // Add service
-        ApiException ex = assertThrows(ApiException.class, () -> handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/1.0", "iamnotarealuser"));
-        assertEquals(LAMBDA_FAILURE, ex.getCode(), "Should have error code 418");
-
-        final long count = testingPostgres.runSelectStatement(
-            "select count(*) from service where sourcecontrol = 'github.com' and organization = 'DockstoreTestUser2' and repository = 'test-service'",
-            long.class);
-        assertEquals(0, count, "there should be no matching service");
+    @Disabled("https://github.com/dockstore/dockstore/pull/4720")
+    void testTRSOutputOfService() {
+        new CreateContent().invoke();
+        final ApiClient webClient = getWebClient(true, false);
+        Ga4GhApi client = new Ga4GhApi(webClient);
+        final List<Tool> tools = client.toolsGet(null, null, null, null, null, null, null, null, null, null, null);
+        assertTrue(tools.stream().filter(tool -> tool.getToolclass().getName().equalsIgnoreCase("workflow")).count() >= 1);
+        // TODO: change boolean once services are exposed
+        boolean servicesExposedInTRS = false;
+        if (servicesExposedInTRS) {
+            assertTrue(tools.stream().filter(tool -> tool.getToolclass().getName().equalsIgnoreCase("service")).count() >= 2);
+        }
     }
 
-    /**
-     * Ensures that a service and workflow can have the same path
-     *
-     */
     @Test
-    void testServiceWithSamePathAsWorkflow() {
-        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
+    void testProprietaryAPI() {
+        final CreateContent invoke = new CreateContent().invoke();
+        final ApiClient webClient = getWebClient(true, false);
         WorkflowsApi client = new WorkflowsApi(webClient);
+        final List<io.swagger.client.model.Workflow> services = client.allPublishedWorkflows(null, null, null, null, null, true, null);
+        final List<io.swagger.client.model.Workflow> workflows = client.allPublishedWorkflows(null, null, null, null, null, false, null);
+        assertTrue(workflows.size() >= 2 && workflows.stream()
+            .noneMatch(workflow -> workflow.getDescriptorType().getValue().equalsIgnoreCase(DescriptorLanguage.SERVICE.toString())));
+        Client jerseyClient = new JerseyClientBuilder(SUPPORT.getEnvironment()).build("test client");
+        CommonTestUtilities.testXTotalCount(jerseyClient, String.format("http://localhost:%d/workflows/published", SUPPORT.getLocalPort()), 2);
+        CommonTestUtilities.testXTotalCount(jerseyClient, String.format("http://localhost:%d/workflows/published?services=true", SUPPORT.getLocalPort()), 2);
+        assertTrue(services.size() >= 1 && services.stream()
+            .allMatch(workflow -> workflow.getDescriptorType().getValue().equalsIgnoreCase(DescriptorLanguage.SERVICE.toString())));
 
-        // Add service
-        handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/1.0", USER_2_USERNAME);
-        long workflowCount = testingPostgres.runSelectStatement("select count(*) from service", long.class);
-        assertEquals(1, workflowCount);
+        // try some standard things we would like services to be able to do
+        client.starEntry(invoke.getServiceID(), new StarRequest().star(true));
+        client.updateLabels(invoke.getServiceID(), "foo,batman,chicken", "");
 
-        // Add workflow with same path as service
-        final Workflow workflow = client
-            .manualRegister("github", DockstoreTestUser2Repos.TEST_SERVICE, "/Dockstore.cwl", "", "cwl", "/test.json");
-        assertNotNull(workflow);
-
-        // forcibly publish both for testing
-        testingPostgres.runUpdateStatement("update workflow set ispublished = 't', waseverpublic = 't'");
-        testingPostgres.runUpdateStatement("update service set ispublished = 't', waseverpublic = 't'");
-
-        // test retrieval
-        final Workflow returnedWorkflow = client.getPublishedWorkflowByPath(SourceControl.GITHUB + "/" + DockstoreTestUser2Repos.TEST_SERVICE, WorkflowSubClass.BIOWORKFLOW, "",  null);
-        final Workflow returnedService = client.getPublishedWorkflowByPath(SourceControl.GITHUB + "/" + DockstoreTestUser2Repos.TEST_SERVICE, WorkflowSubClass.SERVICE, "",  null);
-        assertNotSame(returnedWorkflow.getId(), returnedService.getId());
-
-        // test GA4GH retrieval
-        Ga4GhApi ga4GhApi = new Ga4GhApi(getWebClient(USER_2_USERNAME, testingPostgres));
-        final Tool tool1 = ga4GhApi.toolsIdGet(EntryTypeMetadata.WORKFLOW.getTrsPrefix() + "/" + SourceControl.GITHUB + "/" + DockstoreTestUser2Repos.TEST_SERVICE);
-        final Tool tool2 = ga4GhApi.toolsIdGet(EntryTypeMetadata.SERVICE.getTrsPrefix() + "/" + SourceControl.GITHUB + "/" + DockstoreTestUser2Repos.TEST_SERVICE);
-        assertNotSame(tool1.getId(), tool2.getId());
+        // did it happen?
+        final io.swagger.client.model.Workflow workflow = client.getWorkflow(invoke.getServiceID(), "");
+        assertFalse(workflow.getStarredUsers().isEmpty());
+        assertTrue(workflow.getLabels().stream().anyMatch(label -> "batman".equals(label.getValue())));
     }
 
-    /**
-     * This tests that you can't add a version that doesn't exist
-     */
     @Test
-    void updateServiceIncorrectTag() {
-        final ApiClient webClient = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
+    void testGeneralDefaultPathMechanism() {
+        final CreateContent invoke = new CreateContent().invoke();
+        final ApiClient webClient = getWebClient(true, false);
         WorkflowsApi client = new WorkflowsApi(webClient);
-
-        // Add version that doesn't exist
-        ApiException ex = assertThrows(ApiException.class, () -> handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/1.0-fake", ADMIN_USERNAME));
-        assertEquals(LAMBDA_FAILURE, ex.getCode(), "Should have error code 418");
+        // did it happen?
+        final io.swagger.client.model.Workflow workflow = client.getWorkflow(invoke.getServiceID(), "");
     }
 
-    /**
-     * This tests that you can't add a version with an invalid dockstore.yml or no dockstore.yml
-     */
-    @Test
-    void updateServiceNoOrInvalidYml() {
-        final ApiClient webClient = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
-        WorkflowsApi client = new WorkflowsApi(webClient);
+    private class CreateContent {
+        private long workflowID;
+        private long serviceID;
+        private long serviceID2;
 
-        // Add version that has no dockstore.yml
-        ApiException ex = assertThrows(ApiException.class, () -> handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/no-yml", ADMIN_USERNAME));
-        assertEquals(LAMBDA_FAILURE, ex.getCode(), "Should have error code 418");
+        long getWorkflowID() {
+            return workflowID;
+        }
 
-        // Add version that has invalid dockstore.yml
-        ex = assertThrows(ApiException.class, () -> handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/invalid-yml", ADMIN_USERNAME));
-        assertEquals(LAMBDA_FAILURE, ex.getCode(), "Should have error code 418");
-    }
+        long getServiceID() {
+            return serviceID;
+        }
 
-    /**
-     * Tests that refresh will only grab the releases
-     */
-    @Test
-    void updateServiceSync() {
-        testingPostgres.runUpdateStatement("update enduser set isadmin = 't' where username = 'DockstoreTestUser2';");
-        final ApiClient webClient = getOpenAPIWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi client = new WorkflowsApi(webClient);
+        long getServiceID2() {
+            return serviceID2;
+        }
 
-        // Add service
-        handleGitHubRelease(client, DockstoreTestUser2Repos.TEST_SERVICE, "refs/tags/1.0", USER_2_USERNAME);
-        long workflowCount = testingPostgres.runSelectStatement("select count(*) from service", long.class);
-        assertEquals(1, workflowCount);
+        CreateContent invoke() {
+            return invoke(true);
+        }
 
-        Workflow service = client.getWorkflowByPath("github.com/" + DockstoreTestUser2Repos.TEST_SERVICE, WorkflowSubClass.SERVICE, "");
-        ApiException ex = assertThrows(ApiException.class, () -> client.refresh1(service.getId(), false));
-        assertEquals(HttpStatus.SC_BAD_REQUEST, ex.getCode(), "Should not be able to refresh a dockstore.yml service.");
-    }
+        CreateContent invoke(boolean cleanup) {
+            final Transaction transaction = session.beginTransaction();
 
-    /**
-    * This tests that you cannot create a service from an in invalid GitHub repository
-    */
-    @Test
-    void createServiceNoGitHubRepo() {
-        final ApiClient webClient = getOpenAPIWebClient(ADMIN_USERNAME, testingPostgres);
-        WorkflowsApi client = new WorkflowsApi(webClient);
+            Workflow testWorkflow = new BioWorkflow();
+            testWorkflow.setDescription("foo workflow");
+            testWorkflow.setIsPublished(true);
+            testWorkflow.setSourceControl(SourceControl.GITHUB);
+            testWorkflow.setDescriptorType(DescriptorLanguage.CWL);
+            testWorkflow.setOrganization("shield");
+            testWorkflow.setRepository("shield_repo");
 
-        String serviceRepo = "DockstoreTestUser2/test-service-foo-bar-not-real";
-        ApiException ex = assertThrows(ApiException.class, () -> handleGitHubRelease(client, serviceRepo, "refs/tags/1.0", ADMIN_USERNAME));
-        assertEquals(LAMBDA_FAILURE, ex.getCode(), "Should have error code 418");
+            Service testService = new Service();
+            testService.setDescription("test service");
+            testService.setIsPublished(true);
+            testService.setSourceControl(SourceControl.GITHUB);
+            testService.setDescriptorType(DescriptorLanguage.SERVICE);
+            testService.setMode(WorkflowMode.DOCKSTORE_YML);
+            testService.setOrganization("hydra");
+            testService.setRepository("hydra_repo");
+            testService.setDefaultWorkflowPath(DOCKSTORE_YML_PATH);
+
+            Service test2Service = new Service();
+            test2Service.setDescription("test service");
+            test2Service.setIsPublished(true);
+            test2Service.setSourceControl(SourceControl.GITHUB);
+            test2Service.setMode(WorkflowMode.DOCKSTORE_YML);
+            test2Service.setDescriptorType(DescriptorLanguage.SERVICE);
+            test2Service.setOrganization("hydra");
+            test2Service.setRepository("hydra_repo2");
+            test2Service.setDefaultWorkflowPath(DOCKSTORE_YML_PATH);
+
+            final Map<DescriptorLanguage.FileType, String> defaultPaths = test2Service.getDefaultPaths();
+            for (DescriptorLanguage.FileType val : DescriptorLanguage.FileType.values()) {
+                defaultPaths.put(val, "path for " + val);
+            }
+            test2Service.setDefaultPaths(defaultPaths);
+
+            // add all users to all things for now
+            for (User user : userDAO.findAll()) {
+                testWorkflow.addUser(user);
+                testService.addUser(user);
+                test2Service.addUser(user);
+            }
+
+            workflowID = workflowDAO.create(testWorkflow);
+            serviceID = serviceDAO.create(testService);
+            serviceID2 = serviceDAO.create(test2Service);
+
+            assertTrue(workflowID != 0 && serviceID != 0);
+
+            session.flush();
+            transaction.commit();
+            if (cleanup) {
+                session.close();
+            }
+            return this;
+        }
     }
 }
